refactor(id): migrate from Integer to Long for entity IDs`

+++ common/src/main/java/ru/yandex/practicum/filmorate/common/infrastructure/storage/inmemory/AbstractInMemoryRepository.java
@@ -8,20 +8,20 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 
 @Slf4j
 @RequiredArgsConstructor
 public abstract class AbstractInMemoryRepository<T, C, U> {
-  protected final ConcurrentHashMap<Integer, T> storage = new ConcurrentHashMap<>();
-  private final AtomicInteger idGenerator = new AtomicInteger(0);
+  protected final ConcurrentHashMap<Long, T> storage = new ConcurrentHashMap<>();
+  private final AtomicLong idGenerator = new AtomicLong(0);
 
-  private final BiFunction<Integer, C, T> createBuilder;
+  private final BiFunction<Long, C, T> createBuilder;
   private final Function<U, T> updateBuilder;
-  private final Function<U, Integer> updateIdExtractor;
-  private final Function<T, Integer> entityIdExtractor;
+  private final Function<U, Long> updateIdExtractor;
+  private final Function<T, Long> entityIdExtractor;
 
   public final T save(C createCommand) {
     T value = createBuilder.apply(generateNextId(),
@@ -31,7 +31,7 @@ public abstract class AbstractInMemoryRepository<T, C, U> {
     return value;
   }
 
-  public final Integer generateNextId() {
+  public final Long generateNextId() {
     return idGenerator.incrementAndGet();
   }
 
@@ -40,7 +40,7 @@ public abstract class AbstractInMemoryRepository<T, C, U> {
   }
 
   public final T update(U updateCommand) {
-    Integer id = updateIdExtractor.apply(updateCommand);
+    Long id = updateIdExtractor.apply(updateCommand);
     if (!storage.containsKey(id)) {
       String errorMessage = "Entity with id " + id + " not found";
       log.warn(errorMessage);
@@ -52,7 +52,7 @@ public abstract class AbstractInMemoryRepository<T, C, U> {
     return value;
   }
 
-  public final Optional<T> findById(int id) {
+  public final Optional<T> findById(long id) {
     return Optional.ofNullable(storage.get(id));
   }
 }
