+++ films/src/main/java/ru/yandex/practicum/filmorate/films/infrastructure/web/dto/FilmResponse.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.films.infrastructure.web.dto;
 
 import java.time.LocalDate;
 
-public record FilmResponse(Integer id,
+public record FilmResponse(Long id,
                            String name,
                            String description,
                            LocalDate releaseDate,
